sdfg

+++ cogs/cmd_secret_chats.py
@@ -77,8 +77,9 @@ class secret_chats(commands.Cog):
             for message in all_messages:
                 remove_emoji = 'trg_removing'
                 for reaction in message.reactions:
-                    if remove_emoji in str(reaction):
-                        await message.delete()
+                    async for user in reaction.users:
+                        if user == self.client.user and remove_emoji in str(reaction):
+                            await message.delete()
             print('clean up done')
         except:
             e = sys.exc_info()[0]
